requirements.txt
torch>=2.0.0
diffusers>=0.19.0
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.23.0
requests>=2.28.0
tqdm>=4.65.0
trimesh>=3.22.0
pyrender>=0.1.45
pyglet>=2.0.0
gradio>=3.35.0
argparse>=1.4.0
matplotlib>=3.7.0
scikit-image>=0.20.0
opencv-python>=4.7.0
numpy-stl>=3.0.0
pygltflib>=1.15.0
huggingface-hub>=0.12.0